Better support for view pages
hook_views_pre_render is not called before breadcrumb hooks, so breadcrumbs of view pages were failing.  For now, simply set the context if we don't already know what it is.  This would be better handled by hook_og_context_negotiation_info, but we can't reference that in a template.